Dockerize for development, face management, tests, etc. Lots of goodies!

+++ face_manager/models.py
@@ -6,6 +6,7 @@ from django.core.validators import *
 from datetime import datetime
 from django.utils import timezone
 from django.contrib.postgres.fields import ArrayField
+import cv2
 
 import os
 import sys
@@ -29,6 +30,34 @@ def face_highlight_path(instance, filename):
     first_dir = filename[:1]
     return f"face_highlights/{first_dir}/{filename}"
 
+def get_default_blank_person():
+    
+    # Get the blank face name person, or create if they don't
+    # exist.
+    blank_face_default = Person.objects.get(person_name=settings.BLANK_FACE_NAME)
+    if blank_face_default is None:
+        new_person = Person(person_name = settings.BLANK_FACE_NAME)
+        # Read the blank face image (thanks XKCD). It's saved
+        # in the project now. 
+        person_thumb = cv2.imread(settings.BLANK_FACE_IMG_PATH)
+        personal_thumbnail = cv2.resize(sq_thumb, (500, 500))
+        # encode the image
+        is_success, person_buff = cv2.imencode(".jpg", personal_thumbnail)
+        # Save thumbnail to in-memory file as BytesIO
+        person_byte_thumbnail = BytesIO(person_buff)
+
+        # Fun binary filename and such.
+        person_thumb_filename = f'033003300DEADBEEF003_BLANK.jpg'
+        person_byte_thumbnail.seek(0)
+        # Load a ContentFile into the thumbnail field so it gets saved
+        new_person.highlight_img.save(person_thumb_filename, ContentFile(person_byte_thumbnail.read())) 
+
+        new_person.save()
+        
+        blank_face_default = Person.objects.get(person_name = settings.BLANK_FACE_NAME)
+
+    return blank_face_default
+
 class Person(models.Model):
     person_name = models.CharField(max_length=256)
     # I'd rather the highlight be its own image rather than
@@ -37,6 +66,12 @@ class Person(models.Model):
 
     # id field has a primary key
     def delete(self):
+        # Protect the no face assigned person from
+        # ever being deleted. 
+        if self.person_name == settings.BLANK_FACE_NAME:
+            return
+        # Else, remove the saved image and 
+        # then delete the person object. 
         try:
             os.remove(self.highlight_img.path)
         except ValueError as ve:
@@ -50,8 +85,15 @@ class Person(models.Model):
         return self.person_name
 
 class Face(models.Model):
-    # Primary key comes for free
-    declared_name = models.ForeignKey('Person', on_delete=models.PROTECT, related_name='face_declared', blank=True, null=True)
+    
+    # Primary key (id) comes for free.
+    # For all the foreign keys to person, we set the on_delete method
+    # to models.SET. This property takes a function that returns
+    # a given object -- i.e. the "_NO_FACE_ASSIGNED_" person.
+    # Importantly, the field takes a function rather than
+    # an object. See above for the definition of the function.
+    declared_name = models.ForeignKey('Person', on_delete=models.SET(get_default_blank_person), related_name='face_declared', \
+        blank=True, null=True)
     source_image_file = models.ForeignKey('filepopulator.ImageFile', on_delete=models.CASCADE, blank=True, null=True)
     # ArrayField supported in PostGres
     face_encoding = ArrayField(
@@ -61,20 +103,31 @@ class Face(models.Model):
 
     # This field will contain the top 5 possible identities as categorized
     # by the FC network.
-
-    poss_ident1 = models.ForeignKey('Person', on_delete=models.PROTECT, related_name='face_poss1', blank=True, null=True)
-    poss_ident2 = models.ForeignKey('Person', on_delete=models.PROTECT, related_name='face_poss2', blank=True, null=True)
-    poss_ident3 = models.ForeignKey('Person', on_delete=models.PROTECT, related_name='face_poss3', blank=True, null=True)
-    poss_ident4 = models.ForeignKey('Person', on_delete=models.PROTECT, related_name='face_poss4', blank=True, null=True)
-    poss_ident5 = models.ForeignKey('Person', on_delete=models.PROTECT, related_name='face_poss5', blank=True, null=True)
+    # Like the declared name, these also have the on_delete method set. 
+    # No need to worry much about this -- these fields are transient
+    # by nature and will be reassigned to existing people objects
+    # the next time the FC network runs.
+    poss_ident1 = models.ForeignKey('Person', on_delete=models.SET(get_default_blank_person), related_name='face_poss1', \
+        blank=True, null=True)
+    poss_ident2 = models.ForeignKey('Person', on_delete=models.SET(get_default_blank_person), related_name='face_poss2', \
+        blank=True, null=True)
+    poss_ident3 = models.ForeignKey('Person', on_delete=models.SET(get_default_blank_person), related_name='face_poss3', \
+        blank=True, null=True)
+    poss_ident4 = models.ForeignKey('Person', on_delete=models.SET(get_default_blank_person), related_name='face_poss4', \
+        blank=True, null=True)
+    poss_ident5 = models.ForeignKey('Person', on_delete=models.SET(get_default_blank_person), related_name='face_poss5', \
+        blank=True, null=True)
 
     written_to_photo_metadata = models.BooleanField(default=False)
 
+    # Preserve the values of the face's bounding box.
     box_top = models.IntegerField(validators=[MinValueValidator(1)], default=-1)
     box_bottom = models.IntegerField(validators=[MinValueValidator(1)], default=-1)
     box_left = models.IntegerField(validators=[MinValueValidator(1)], default=-1)
     box_right = models.IntegerField(validators=[MinValueValidator(1)], default=-1)
 
+    # A field to save the thumbnail. The scripts.py ensures
+    # that this is a square thumbnail.
     face_thumbnail = models.ImageField(upload_to=face_thumbnail_path, default=None)
